Move CLI-level support bundle tests into test_cli.py per plan and delete stray test_support_bundle.py
